02_SSBW/05_PROYECTO/senderos/requirements.txt
Django==3.2
django-crispy-forms
psycopg2-binary>=2.8
mongoengine
django-allauth
djangorestframework>=3.1,<4.0
blinker==1.4
djangorestframework-simplejwt>=4.6,<5.0
django-cors-headers>=3.7, <4.0
Pillow==8.1.1
sorl-thumbnail
djongo
sqlparse==0.2.4
uuid